iox-#2278 Split string tests into multiple files
